settings: rename cssNight.js to cssDark.js
